Added tee to dry run; tmp file created by mktemp

+++ lsfTophat.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
-set -e
+set -e -o pipefail
+# Tommaso Leonardi, tl344@ebi.ac.uk
+# This script allows to run Tophat on an LSF system with a shared filesystem
+# It runs Tophat on a node using a temporary local folder to store output.
+# Upon completion, it copies the output back to a user-specified path
 
 usage()
 {
@@ -20,9 +24,6 @@ OPTIONS:
 
 EOF
 }
-# For some unknown reason this script failed to copy back the results folders
-# To get them back I've used:
-# for i in *.log; do HOST=$(grep "Job was executed on host" $i | perl -pe 's/Job was executed on host\(s\) \<(.+)\>, in queue .*/$1/'); FOLDER=$(grep "A summary of the alignment counts can be found in" $i | perl -pe 's/.+A summary of the alignment counts can be found in (\/.+)align.summary.txt$/$1/'); JOB=$(grep "was submitted from host" $i | perl -pe 's/Job \<(.+)\> was submitted from host .+/$1/'); echo $HOST : $FOLDER : $JOB;mkdir -p $BASEDIR/transcriptomes/tophat/$JOB; scp $HOST:$FOLDER/* $BASEDIR/transcriptomes/tophat/$JOB;done
 
 while getopts “hp:o:i:f:t:d” OPTION
 do
@@ -57,15 +58,20 @@ if [ -z "$PARAMS" ] || [ -z "$OUT" ] || [ -z "$INDEX" ] || [ -z "$FASTQ" ] || [ 
 	exit 1;
 fi
 
-TMP=/tmp/$RANDOM$RANDOM
-mkdir $TMP
+if [ -d "$OUT" ]; then
+	echo "Error: Output directory $OUT already exists";
+	exit 1;
+fi
+mkdir -p $OUT
+
+TMP=$(mktemp -d --tmpdir=/tmp)
 
 if [ -z "$DRY" ]; then
 	$TOPHAT $PARAMS -o $TMP $INDEX $FASTQ;
 else
-	echo $TOPHAT $PARAMS -o $TMP $INDEX $FASTQ > $TMP/test_tophat.output;
+	echo "$TOPHAT $PARAMS -o $TMP $INDEX $FASTQ" | tee $TMP/test_tophat.output;
 fi
-mkdir -p $OUT
-mv $TMP/* $OUT
+
+cp -R $TMP/* $OUT
 rm -rf $TMP
